Made NewListDialog its own component
